requirements.txt
adafruit-board-toolkit
pyserial